improves initialization system so user can instatiate multiple objects without risk of crash, adds further crash protection in imperfect usage and system failure
