+++ CMakeLists.txt
@@ -7,8 +7,14 @@ IF("${DEBUG_SWITCH}" STREQUAL "DEBUG_UNIX")
 	ADD_DEFINITIONS(-DDEBUG_UNIX)
 	MESSAGE(STATUS "Debugging UNIX")
 ELSE("${DEBUG_SWITCH}" STREQUAL "DEBUG_UNIX")
-	ADD_DEFINITIONS(-DDEBUG_WINDOWS)
-	MESSAGE(STATUS "Debugging WINDOWS")
+	IF("${DEBUG_SWITCH}" STREQUAL "DEBUG_UNIX_ON_WINDOWS")
+		ADD_DEFINITIONS(-DDEBUG_UNIX_ON_WINDOWS)
+		SET(UX TRUE)
+		MESSAGE(STATUS "Debugging UNIX on WINDOWS")
+	ELSE("${DEBUG_SWITCH}" STREQUAL "DEBUG_UNIX_ON_WINDOWS")
+		ADD_DEFINITIONS(-DDEBUG_WINDOWS)
+		MESSAGE(STATUS "Debugging WINDOWS")
+	ENDIF("${DEBUG_SWITCH}" STREQUAL "DEBUG_UNIX_ON_WINDOWS")
 ENDIF("${DEBUG_SWITCH}" STREQUAL "DEBUG_UNIX")
 
 IF("${CMAKE_BUILD_TYPE}" STREQUAL "NO_MEX")
@@ -25,7 +31,7 @@ IF("${CMAKE_BUILD_TYPE}" STREQUAL "NO_MEX")
 
 	SET(SOURCE_FILES
 			src/matshare_.c src/headers/matshare_.h
-		   src/utils.c src/headers/utils.h)
+		   src/utils.c src/headers/utils.h src/init.c src/headers/init.h src/headers/msh_types.h)
 
 	ADD_EXECUTABLE(${OUT_NAME} ${SOURCE_FILES})
