+++ bin/msh_fetch.m
@@ -1,0 +1,5 @@
+function out = msh_fetch
+	% get callback
+	out = matshare_(uint8(5));
+end
+
