+++ bin/msh_free.m
@@ -1,0 +1,4 @@
+function msh_free
+	matshare_(uint8(4));
+end
+
