+++ bin/msh_share.m
@@ -1,0 +1,5 @@
+function msh_share(in)
+	% set fcn
+	matshare_(uint8(1), in);
+end
+
